Add YAML-based configuration and modularize the code
This change adds the r-yaml dependency and references it in
DESCRIPTION, and properly registers modelc as a dependency

The r-yaml dependency was added so that we could remove the previous approach to configuration
which relied on global variables and replaces it with a YAML based
configuration that is local to the caller functions.

This change also removes the main.r module and introduces the db.r,
update.r, utils.r, and interface.r modules.

db.r contains a reference class for managing the lifecycle of a
database and encapsulating database-related functionality. This code
can be further decomposed into modules responsible for generating SQL
and for performing database operations. Extra support was added for
cleaning up database connections

update.r contains the function which executes the main loop in the
model upsert logic.

interface.r contains functions that are a part of the
user-interface. This code can be further decomposed into modules
responsible for the visual elements of the interface and the interface
functions themselves.